+++ MyRADProject/Assets/RobsCharacter.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.44657898, g: 0.4964133, b: 0.5748178, a: 1}
+  m_IndirectSpecularColor: {r: 0.44657874, g: 0.49641275, b: 0.5748172, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -20529,6 +20529,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 6cd866ce660af124a814fb1bfad36b9f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  boxSize: {x: 0, y: 0, z: 0}
+  maxDistance: 3
+  layerMask:
+    serializedVersion: 2
+    m_Bits: 0
+  jumpforce: 5
+  snowballCloneTemplate: {fileID: 6598988467822025660, guid: b45bb0a0538aed642bf6629b5f398cb0, type: 3}
 --- !u!114 &2117685198
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -20541,10 +20548,10 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 0756b311449205c44918d6daa3ab3a6b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  rb: {fileID: 0}
   boxSize: {x: 0, y: 0, z: 0}
   maxDistance: 0.7
   layerMask:
-    serializedVersion: 2
     m_Bits: 0
   jumpforce: 4
 --- !u!4 &2117685202 stripped
@@ -20560,7 +20567,7 @@ VFXRenderer:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2117685187}
-  m_Enabled: 1
+  m_Enabled: 0
   m_CastShadows: 0
   m_ReceiveShadows: 0
   m_DynamicOccludee: 1
@@ -20599,7 +20606,7 @@ VisualEffect:
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2117685187}
-  m_Enabled: 1
+  m_Enabled: 0
   m_Asset: {fileID: 0}
   m_InitialEventName: OnPlay
   m_InitialEventNameOverriden: 0
