+++ MyRADProject/Assets/Snowball.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: ae227a881cede1a4d89c06c8643dc95a
-NativeFormatImporter:
+guid: b45bb0a0538aed642bf6629b5f398cb0
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 15600000
   userData: 
   assetBundleName: 
   assetBundleVariant:
