+++ MyRADProject/Assets/Tree 1_Textures/Tree.prefab.meta
